Make get result of type double

+++ App1/App1/ICalculator.cs
@@ -6,6 +6,6 @@ namespace App1
 {
     public interface ICalculator
     {
-        int GetResult();
+        double GetResult();
     }
 }
